// ChargeMefront/src/app/models/born.ts
export class Born {
    id?: number;
    name?: string;
    serialNumber?: string;
    etatB?: string;
    proprietaire?: string;
    pourcentageP?: string;
    NumeroTel?: string;
    createdAt?: string;
    updatedAt?: string;
    StationId?: number;
    prixParKilowatt?: number;
    disponibilite?: string;
    emplacement?: string;
    modele?:string;
    image?: string;
}